+++ GANDLF/data/ImagesFromDataFrame.py
@@ -9,24 +9,9 @@ from torchio.transforms import (OneOf, RandomMotion, RandomGhosting, RandomSpike
 from torchio import Image, Subject
 import SimpleITK as sitk
 from GANDLF.utils import resize_image
+from GANDLF.preprocessing import *
 
 import copy
-
-def threshold_intensities(input_tensor, min_val, max_val):
-    '''
-    This function takes an input tensor and 2 thresholds, lower & upper and thresholds between them, basically making intensity values outside this range '0'
-    '''
-    l1_tensor = torch.where(input_tensor < min_val, input_tensor, 0)
-    l2_tensor = torch.where(l1_tensor > max_val, l1_tensor, 0)
-    return l2_tensor
-
-def clip_intensities(input_tensor, min_val, max_val):
-    '''
-    This function takes an input tensor and 2 thresholds, lower and upper and clips between them, basically making the lowest value as 'min_val' and largest values as 'max_val'
-    '''
-    l1_tensor = torch.where(input_tensor < min_val, input_tensor, min_val)
-    l2_tensor = torch.where(l1_tensor > max_val, l1_tensor, max_val)
-    return l2_tensor
 
 ## todo: ability to change interpolation type from config file
 ## todo: ability to change the dimensionality according to the config file
@@ -46,12 +31,6 @@ def spatial_transform(patch_size = None, p=1):
         max_displacement = 7.5
     return OneOf({RandomAffine(): 0.8, RandomElasticDeformation(max_displacement = max_displacement): 0.2}, p=p)
 
-def threshold_transform(min, max, p=1):
-    return Lambda(lambda x: threshold_intensities(x, min, max))
-
-def clip_transform(min, max, p=1):
-    return Lambda(lambda x: clip_intensities(x, min, max))
-
 def bias(patch_size = None, p=1):
     return RandomBiasField(coefficients=0.5, order=3, p=p, seed=None)
 
@@ -64,11 +43,22 @@ def noise(patch_size = None, p=1):
 def swap(patch_size = 15, p=1):
     return RandomSwap(patch_size=patch_size, num_iterations=100, p=p, seed=None)
 
-# Defining a dictionary - key is the string and the value is the augmentation object
-global_augs_dict = {
+## lambdas for pre-processing
+def threshold_transform(min, max, p=1):
+    return Lambda(lambda x: threshold_intensities(x, min, max))
+
+def clip_transform(min, max, p=1):
+    return Lambda(lambda x: clip_intensities(x, min, max))
+
+# defining dict for pre-processing - key is the string and the value is the transform object
+global_preprocessing_dict = {
     'threshold' : threshold_transform,
     'clip' : clip_transform,
-    'normalize' : ZNormalization(),
+    'normalize' : ZNormalization()
+}
+
+# Defining a dictionary for augmentations - key is the string and the value is the augmentation object
+global_augs_dict = {
     'spatial' : spatial_transform,
     'kspace' : mri_artifact,
     'bias' : bias,
@@ -79,7 +69,7 @@ global_augs_dict = {
 
 # This function takes in a dataframe, with some other parameters and returns the dataloader
 def ImagesFromDataFrame(dataframe, psize, headers, q_max_length, q_samples_per_volume,
-                        q_num_workers, q_verbose, train=True, augmentations=None, resize=None):
+                        q_num_workers, q_verbose, train=True, augmentations=None, preprocessing=None, resize=None):
     # Finding the dimension of the dataframe for computational purposes later
     num_row, num_col = dataframe.shape
     # num_channels = num_col - 1 # for non-segmentation tasks, this might be different
@@ -98,6 +88,7 @@ def ImagesFromDataFrame(dataframe, psize, headers, q_max_length, q_samples_per_v
         augmentation_patchAxesPoints[i] = max(round(augmentation_patchAxesPoints[i] / 10), 1) # always at least have 1
 
     # iterating through the dataframe
+    resizeCheck = False
     for patient in range(num_row):
         # We need this dict for storing the meta data for each subject
         # such as different image modalities, labels, any other data
@@ -109,23 +100,23 @@ def ImagesFromDataFrame(dataframe, psize, headers, q_max_length, q_samples_per_v
             subject_dict[str(channel)] = Image(str(dataframe[channel][patient]),
                                                type=torchio.INTENSITY)
 
-            if resize is not None:
-                image_resized = resize_image(subject_dict[str(channel)].as_sitk(), resize)
-                image_masked_tensor = torch.from_numpy(np.swapaxes(sitk.GetArrayFromImage(image_resized), 0, 2))
-                # overwrite previous image data with new masked data
-                subject_dict[str(channel)] = Image(tensor=image_masked_tensor,
-                                                   type=torchio.INTENSITY)
+            # if resize has been defined but resample is not (or is none)
+            if not resizeCheck:
+                if (preprocessing['resize'] is not None):
+                    resizeCheck = True
+                    if not('resample' in preprocessing):
+                        preprocessing['resample'] = {}
+
+                        if not('resolution' in preprocessing['resample']):
+                            preprocessing['resample']['resolution'] = resize_image_resolution(subject_dict[str(channel)].as_sitk(), resize)
+                    else:
+                        print('WARNING: \'resize\' is ignored as \'resample\' is defined under \'data_processing\', this will be skipped', file = sys.stderr)
+                else:
+                    resizeCheck = True
+
 
         if labelHeader is not None:
             subject_dict['label'] = Image(str(dataframe[labelHeader][patient]), type=torchio.LABEL)
-
-            if resize is not None:
-                image_resized = resize_image(subject_dict['label'].as_sitk(),
-                                             resize,
-                                             sitk.sitkNearestNeighbor)
-                image_masked_tensor = torch.from_numpy(np.swapaxes(sitk.GetArrayFromImage(image_resized), 0, 2))
-                # overwrite previous image data with new masked data
-                subject_dict['label'] = Image(tensor = image_masked_tensor, type=torchio.INTENSITY)
 
             if not train:
                 subject_dict['path_to_metadata'] = str(dataframe[labelHeader][patient])
@@ -142,28 +133,27 @@ def ImagesFromDataFrame(dataframe, psize, headers, q_max_length, q_samples_per_v
 
     # first, we want to do thresholding, followed by clipping, if it is present - required for inference as well
     for key in ['threshold','clip']:
-        if key in augmentations:
-            augmentation_list.append(global_augs_dict[key](min=augmentations[key]['min'], max=augmentations[key]['max']))
+        augmentation_list.append(global_preprocessing_dict[key](min=preprocessing[key]['min'], max=preprocessing[key]['max']))
         
     # first, we want to do the resampling, if it is present - required for inference as well
-    if 'resample' in augmentations:
-        if 'resolution' in augmentations['resample']:
+    if 'resample' in preprocessing:
+        if 'resolution' in preprocessing['resample']:
             # resample_split = str(aug).split(':')
-            resample_values = tuple(np.array(augmentations['resample']['resolution']).astype(np.float))
+            resample_values = tuple(np.array(preprocessing['resample']['resolution']).astype(np.float))
+            if len(resample_values) == 2:
+                resample_values = np.append(resample_values,1)
             augmentation_list.append(Resample(resample_values))
 
     # next, we want to do the intensity normalize - required for inference as well
-    if 'normalize' in augmentations:
-        augmentation_list.append(global_augs_dict['normalize'])
+    if 'normalize' in preprocessing:
+        augmentation_list.append(global_preprocessing_dict['normalize'])
 
     # other augmentations should only happen for training - and also setting the probabilities
     # for the augmentations
-    if train:
+    if train and not(augmentations == None):
         for aug in augmentations:
-            # resample and normalize should always have probability=1
-            if (aug != 'normalize') and (aug != 'resample'):
-                actual_function = global_augs_dict[aug](patch_size=augmentation_patchAxesPoints, p=augmentations[aug]['probability'])
-                augmentation_list.append(actual_function)
+            actual_function = global_augs_dict[aug](patch_size=augmentation_patchAxesPoints, p=augmentations[aug]['probability'])
+            augmentation_list.append(actual_function)
 
     transform = Compose(augmentation_list)