+++ GANDLF/parseConfig.py
@@ -15,6 +15,18 @@ def parse_version(version_string):
     del version_string_split[-1]
   return int(''.join(version_string_split))
 
+def initialize_key(parameters, key):
+  '''
+  This function will initialize the key in the parameters dict to 'None' if it is absent or length is zero
+  '''
+  if key in parameters: 
+    if len(parameters[key]) == 0: # if key is present but not defined
+      parameters[key] = None
+  else:
+    parameters[key] = None # if key is absent
+
+  return parameters
+
 def parseConfig(config_file_path):
   '''
   This function parses the configuration file and returns a dictionary of parameters
@@ -45,10 +57,7 @@ def parseConfig(config_file_path):
     sys.exit('The \'patch_size\' parameter needs to be present in the configuration file')
   
   if 'resize' in params:
-    if not np.greater_equal(params['resize'], params['psize']).all():
-      sys.exit('The \'resize\' parameter needs to be greater than or equal to \'patch_size\'')
-  else:
-    params['resize'] = None
+    print('WARNING: \'resize\' should be defined under \'data_processing\', this will be skipped', file = sys.stderr)
 
   # Extrating the training parameters from the dictionary
   if 'num_epochs' in params:
@@ -123,28 +132,46 @@ def parseConfig(config_file_path):
   params['opt'] = opt
   
   # this is NOT a required parameter - a user should be able to train with NO augmentations
-  if len(params['data_augmentation']) > 0: # only when augmentations are defined
-    thresholdOrClip = False
-    for key in params['data_augmentation']: # iterate through all keys
-      # for threshold or clip, ensure min and max are defined
-      if not thresholdOrClip:
-        if (key == 'threshold') or (key == 'clip'):
-          thresholdOrClip = True
-          if not(isinstance(params['data_augmentation'][key], dict)):
-            params['data_augmentation'][key] = {}
-          
-          if not 'min' in params['data_augmentation'][key]: 
-            params['data_augmentation'][key]['min'] = sys.float_info.min
-          if not 'max' in params['data_augmentation'][key]:
-            params['data_augmentation'][key]['max'] = sys.float_info.max
-      else:
-        sys.exit('Use only \'threshold\' or \'clip\', not both')
+  params = initialize_key(params, 'data_augmentation')
+  if not(params['data_augmentation'] == None):
+    if len(params['data_augmentation']) > 0: # only when augmentations are defined
+        keysToSkip = ['normalize', 'resample', 'threshold', 'clip']
+        if not(key in keysToSkip): # no need to check probabilities for these: they should ALWAYS be added
+          if (params['data_augmentation'][key] == None) or not('probability' in params['data_augmentation'][key]): # when probability is not present for an augmentation, default to '1'
+              params['data_augmentation'][key] = {}
+              params['data_augmentation'][key]['probability'] = 1
+        else:
+          print('WARNING: \'' + key + '\' should be defined under \'data_processing\' and not under \'data_augmentation\', this will be skipped', file = sys.stderr)
 
-      keysToSkip = ['normalize', 'resample', 'threshold', 'clip']
-      if not(key in keysToSkip): # no need to check probabilities for these: they should ALWAYS be added
-        if (params['data_augmentation'][key] == None) or not('probability' in params['data_augmentation'][key]): # when probability is not present for an augmentation, default to '1'
-            params['data_augmentation'][key] = {}
-            params['data_augmentation'][key]['probability'] = 1
+  # this is NOT a required parameter - a user should be able to train with NO built-in pre-processing 
+  params = initialize_key(params, 'data_preprocessing')
+  if not(params['data_preprocessing'] == None):
+    if len(params['data_preprocessing']) < 0: # perform this only when pre-processing is defined
+      thresholdOrClip = False
+      thresholdOrClipDict = ['threshold', 'clip'] # this can be extended, as required
+      keysForWarning = ['resize'] # properties for which the user will see a warning
+
+      # iterate through all keys
+      for key in params['data_preprocessing']: # iterate through all keys
+        # for threshold or clip, ensure min and max are defined
+        if not thresholdOrClip:
+          if (key in thresholdOrClipDict):
+            thresholdOrClip = True # we only allow one of threshold or clip to occur and not both
+            if not(isinstance(params['data_preprocessing'][key], dict)): # initialize if nothing is present
+              params['data_preprocessing'][key] = {}
+            
+            # if one of the required parameters is not present, initialize with lowest/highest possible values
+            # this ensures the absence of a field doesn't affect processing
+            if not 'min' in params['data_preprocessing'][key]: 
+              params['data_preprocessing'][key]['min'] = sys.float_info.min
+            if not 'max' in params['data_preprocessing'][key]:
+              params['data_preprocessing'][key]['max'] = sys.float_info.max
+        else:
+          sys.exit('Use only \'threshold\' or \'clip\', not both')
+
+        # give a warning for resize
+        if key in keysForWarning:
+          print('WARNING: \'' + key + '\' is generally not recommended, as it changes image properties in unexpected ways.', file = sys.stderr)
 
   # Extracting the model parameters from the dictionary
   if 'base_filters' in params: