+++ HISTORY.md
@@ -1,7 +1,8 @@
 ## 0.0.5
 
 - 2D support added
-- Added option to threshold or clip the input image
+- Pre-processing module added
+  - Added option to threshold or clip the input image
 
 ## 0.0.4